backend/requirements.txt
# =============================================================================
# DevPilot Backend — Python Dependencies
# Tested with Python 3.13
# =============================================================================

# Web framework
fastapi==0.115.12
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.14.1

# Auth
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.0

# File uploads
python-multipart==0.0.20
python-magic-bin==0.4.14

# AI / LLM
langchain==0.3.23
langchain-openai==0.3.11
langchain-google-genai==2.1.12
openai==1.72.0

# File processing
pdfplumber==0.11.6
python-docx==1.1.2
openpyxl==3.1.5

# Object storage
boto3==1.37.33

# Task queue
celery[redis]==5.4.0
redis==5.2.1

# Validation
pydantic[email]==2.11.1
pydantic-settings==2.8.1

# Utilities
python-dotenv==1.1.0
httpx==0.28.1
aiofiles==24.1.0
